backend/requirements.txt
django==5.2.12
djangorestframework==3.16.1
django-cors-headers==4.9.0
fastapi==0.116.1
pymongo==4.16.0
PyJWT==2.10.1
bcrypt==5.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
google-generativeai==0.8.5
uvicorn==0.35.0
